chore: relocate test folder
