+++ plugins/local-storage/tests/__fixtures__/pkg.js
