+++ plugins/local-storage/tests/__fixtures__/readme-test/test-readme-0.0.0.tgz
