+++ plugins/local-storage/tests/__mocks__/Logger.ts
